% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegressionPlot.R
\name{createRegressionPlot}
\alias{createRegressionPlot}
\title{Plot data as a time series and scatter plot comparing a variable from
two instruments or processing level.}
\usage{
createRegressionPlot(varData_units, tickBreaks = NA, dateLabelFormat = NA)
}
\arguments{
\item{varData_units}{list obtained using getLUMARegressionData(). see ?getLUMARegressionData}

\item{tickBreaks}{How often ticks should appear on time series plots. e.g. 6 hours.
See https://www.rdocumentation.org/packages/ggplot2/versions/1.0.0/topics/scale_x_datetime
date_breaks for valid options.}

\item{dateLabelFormat}{The format of date labels on time series plot
e.g. with Hour and Minute separated by : . See date_labels in link above for valid options.}
}
\value{
a scatter plot and time series of the variable.
}
\description{
Plot data as a time series and scatter plot comparing a variable from
two instruments or processing level.
}
