% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLUMAdata.R
\name{getLUMAdata}
\alias{getLUMAdata}
\title{Extract raw or processed LUMA data from cluster.}
\usage{
getLUMAdata(
  instrument,
  level,
  startDate,
  endDate,
  variables,
  fileTimeRes = NA,
  sep = NA,
  variableColNos = NA,
  timeColFormat = NA,
  skipRows = 0
)
}
\arguments{
\item{instrument}{A list containing the instrument name an site in the form
list(id = "INSTID", site = "SITENAME"). INSTID and SITENAME should be
consistent with the LUMA metadata system. Alternitavely you can use the serial
number by list(serial = "serialNo").}

\item{level}{The quality level of the data. Either "RAW" or an integer.}

\item{startDate, endDate}{Start and end date of plot. Should be of class date.}

\item{variables}{A string or vector of strings of the variables to plot.
These must be consisted with variables on LUMA metadata website.}

\item{fileTimeRes}{Time resolution of file this should be consistent with file
names on the cluster e.g. 1min. Only needed when level is not RAW.}

\item{sep}{If data is RAW then what is data separated by e.g. ','.}

\item{variableColNos}{If data is RAW then a vector of the position of the variable
to be plotted.}

\item{timeColFormat}{The format of time columns. A string or vector of strings
if multiple time columns.}

\item{skipRows}{How many rows to skip when reading raw file. Default is none.}

\item{tickBreaks}{How often ticks should appear on plots. e.g. 6 hours.
See https://www.rdocumentation.org/packages/ggplot2/versions/1.0.0/topics/scale_x_datetime
date_breaks for valid options.}

\item{dateLabelFormat}{The format of date labels e.g. %H:%M to have time labels
with Hour and Minute separated by : . See date_labels in link above for valid
options.}
}
\value{
A list of containing the requested data and the variable units.
}
\description{
Extract raw or processed LUMA data from cluster.
}
