% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLUMARegressionData.R
\name{getLUMARegressionData}
\alias{getLUMARegressionData}
\title{Get LUMA data that can be used}
\usage{
getLUMARegressionData(
  instrument1,
  instrument2,
  startDate,
  endDate,
  variable,
  fileTimeRes
)
}
\arguments{
\item{instrument1, instrument2}{the first instruments to compare. These must be lists
containing at least id, site and level. ECpack is an optional name in the list. id
must be consistent with the LUMA metadata site.}

\item{startDate, endDate}{the time span to get data for. Must be class date
or POSIXct.}

\item{variable}{the variable to compare the instruments. Must be consistent with
LUMA metadata site.}

\item{fileTimeRes}{The file time resolution- this must be the same for each instrument.}
}
\value{

}
\description{
Get LUMA data that can be used
}
