% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSitePositionKML.R
\name{makeSitePositionKML}
\alias{makeSitePositionKML}
\title{Create a kml file displaying the locations of an instrument at a given date.}
\usage{
makeSitePositionKML(
  instrumentID,
  layerName,
  DATE = Sys.Date(),
  outDir = getwd()
)
}
\arguments{
\item{instrumentID}{instrument id that must be consistent with LUMA metadata site.}

\item{layerName}{the name of the output file.}

\item{DATE}{the date to use. Must be class 'Date'.}

\item{outDir}{the directory to save the kml file.}
}
\value{
A kml file that can be viewed in google earth.
}
\description{
Create a kml file displaying the locations of an instrument at a given date.
}
\examples{
makeSitePositionKML('CL31', 'ceilometerPositions', as.Date('2020-10-03'), 'ceilkml')
}
