% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Create a time series from a dataframe or output equivalent to that from getLUMAdata.}
\usage{
plotTimeSeries(
  allData,
  tickBreaks = NA,
  dateLabelFormat = NA,
  title = NA,
  SAVEplot = FALSE,
  SAVEname = NA,
  SAVEpath = NULL,
  SAVEsize = NA
)
}
\arguments{
\item{allData}{The data to plot in the form output by getLUMAdata().}

\item{tickBreaks}{How often ticks should appear on plots. e.g. 6 hours.
See https://www.rdocumentation.org/packages/ggplot2/versions/1.0.0/topics/scale_x_datetime
date_breaks for valid options.}

\item{dateLabelFormat}{The format of date labels e.g. %H:%M to have time labels
with Hour and Minute separated by : . See date_labels in link above for valid
options.}

\item{title}{The title of the plot. If not specified then a default will be
generated based on information given.}

\item{SAVEplot}{Boolean whether to save the plot.}

\item{SAVEname}{The name of the plot if SAVEplot = TRUE}

\item{SAVEpath}{The path to folder in which plot should be saved. Leave as null
to save in current working directory.}

\item{SAVEsize}{The size the save the image. This is a 3D vector consisting of
h (height), w (width) and unit. See https://ggplot2.tidyverse.org/reference/ggsave.html
for valid units.}
}
\value{
A time series for each variable.
}
\description{
Create a time series from a dataframe or output equivalent to that from getLUMAdata.
}
