% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSitePositionKML.R
\name{makeSitePositionKML}
\alias{makeSitePositionKML}
\title{Title}
\usage{
makeSitePositionKML(
  instrumentID,
  layerName,
  DATE = Sys.Date(),
  outDir = getwd()
)
}
\arguments{
\item{outDir}{}
}
\value{

}
\description{
Title
}
