% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLUMARegression.R
\name{plotLUMARegression}
\alias{plotLUMARegression}
\title{Title}
\usage{
plotLUMARegression(
  instrument1,
  instrument2,
  startDate,
  endDate,
  variable,
  fileTimeRes,
  tickBreaks = NA,
  dateLabelFormat = NA
)
}
\arguments{
\item{dateLabelFormat}{}
}
\value{

}
\description{
Title
}
