% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Create a time series from a dataframe or output equivalent to that from getLUMAdata.}
\usage{
plotTimeSeries(
  allData,
  tickBreaks = NA,
  dateLabelFormat = NA,
  title = NA,
  SAVEplot = FALSE,
  SAVEname = NA,
  SAVEpath = NULL,
  SAVEsize = NA
)
}
\arguments{
\item{title}{The title of the plot. If not specified then a default will be
generated based on information given.}

\item{SAVEplot}{Boolean whether to save the plot.}

\item{SAVEname}{The name of the plot if SAVEplot = TRUE}

\item{SAVEpath}{The path to folder in which plot should be saved. Leave as null
to save in current working directory.}

\item{SAVEsize}{The size the save the image. This is a 3D vector consisting of
h (height), w (width) and unit. See https://ggplot2.tidyverse.org/reference/ggsave.html
for valid units.}
}
\value{

}
\description{
Create a time series from a dataframe or output equivalent to that from getLUMAdata.
}
